moving all requestor tests in one file
